Move pyproject functions into their own file
